+++ src/ShellKrypt.Desktop/Shell/Dialogs/ModalShell.axaml.cs
@@ -3,7 +3,7 @@ using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace ShellKrypt.UI.Shared.Controls;
+namespace ShellKrypt.Desktop.Shell.Dialogs;
 
 public enum ModalShellSize
 {